Create layers
I'm not sure if this is really ready to push as a post yet. Right now it reads like a restating of the home page. Maybe the next step is to flesh out each of the layers into a full post...  anyway im gonna commit just so you can read and let me know if you have feedback ...